+++ CHANGES.rst
@@ -24,6 +24,9 @@ Specviz
 API Changes
 -----------
 
+- CLI now takes the layout as a required first positional argument after jdaviz
+  (``jdaviz cubeviz path/to/file``). [#1252]
+
 Cubeviz
 ^^^^^^^